python_data_wrangling/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
anyio==4.9.0
    # via httpx
black==25.1.0
    # via -r requirements.in
certifi==2025.1.31
    # via
    #   -r requirements.in
    #   httpcore
    #   httpx
click==8.2.0
    # via black
docopt==0.6.2
    # via -r requirements.in
duckdb==1.2.1
    # via -r requirements.in
faker==37.1.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   -r requirements.in
    #   httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via -r requirements.in
idna==3.7
    # via
    #   -r requirements.in
    #   anyio
    #   httpx
mypy-extensions==1.1.0
    # via black
packaging==25.0
    # via black
pathspec==0.12.1
    # via black
platformdirs==4.3.8
    # via black
psutil==7.0.0
    # via -r requirements.in
python-dotenv==1.1.0
    # via -r requirements.in
redis==5.2.1
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
typing-extensions==4.13.2
    # via anyio
tzdata==2025.2
    # via faker
